More size-cutoff saving bugs
Previously, saved as float which interfered with how C++ used it in an equation with other integers. Completely converted the function estimating a size cutoff to one that just counts the reads.

+++ assembler/src/common/pipeline/config_struct.hpp
@@ -637,7 +637,7 @@ struct debruijn_config {
     bool need_mapping;
 
     int search_distance;
-    double size_cutoff;
+    int size_cutoff;
 
     debruijn_config() :
             use_single_reads(false) {
